pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-infrastructure-infrastructure-base"
version = "0.0.1a1"
description = "Infrastructure layer for pythoneda-artifact/infrastructure-base"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-infrastructure/infrastructure-base"
packages = [{include = "pythonedaartifactinfrastructureinfrastructurebase"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
dulwich = "^0.21.5"
grpcio = "^1.41.0"
paramiko = "^2.11.0"
python = "^3.8"
pythoneda-artifact-event-git-tagging = "^0.0.1a1"
pythoneda-artifact-event-infrastructure-git-tagging = "^0.0.1a1"
pythoneda-artifact-infrastructure-base = "^0.0.1a1"
pythoneda-base = "^0.0.1a15"
pythoneda-infrastructure-base = "^0.0.1a12"
pythoneda-shared-git = "^0.0.1a3"
requests = "^2.28.1"
semver = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
